move tests to fuels
