+++ packages/fuels-signers/Cargo.toml
@@ -33,7 +33,6 @@ tokio = { version = "1.10.1", features = ["full"] }
 tai64 = { version = "4.0", features = ["serde"] }
 
 [dev-dependencies]
-fuels-test-helpers = { path = "../fuels-test-helpers", default-features = false }
 hex = { version = "0.4.3", default-features = false, features = ["std"] }
 tempfile = "3.3.0"
 
@@ -41,7 +40,5 @@ tempfile = "3.3.0"
 default = ["std"]
 std = [
     "fuels-core/std",
-    "fuels-test-helpers/std",
     "fuels-types/std",
 ]
-test-helpers = ["fuel-core"]
